requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
av==16.1.0
certifi==2026.2.25
click==8.3.1
coloredlogs==15.0.1
ctranslate2==4.7.1
exceptiongroup==1.3.1
faster-whisper==1.2.1
filelock==3.25.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.6.0
humanfriendly==10.0
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
numpy==2.2.6
onnxruntime==1.23.2
packaging==26.0
propcache==0.4.1
protobuf==7.34.0
Pygments==2.19.2
python-telegram-bot==22.6
PyYAML==6.0.3
rich==14.3.3
shellingham==1.5.4
sympy==1.14.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
yarl==1.23.0
python-dotenv==1.2.2